Rename _snapshot_runner to _snapshot_format
